+++ src/coordination/dynamic_schedule.rs
@@ -1,32 +1,41 @@
 use crate::coordination::EvaluationTask;
 use crate::{Time, Value};
 use priority_queue::PriorityQueue;
-use rtlola_frontend::mir::{OutputReference, Schedule};
+use rtlola_frontend::mir::{OutputReference, Stream};
+use rtlola_frontend::RtLolaMir;
 use std::cmp::Reverse;
 use std::time::Duration;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 /// A struct representing a scheduled task
-pub(crate) struct ScheduledTask<'a> {
+pub(crate) struct ScheduledTask {
     /// The task to be executed
-    task: EvaluationTask<'a>,
+    task: EvaluationTask,
     /// The period of the task. Used to reschedule this task when popped.
     period: Duration,
 }
 
 #[derive(Debug, Clone)]
-pub(crate) struct ScheduleSlice<'a>(Vec<(Time, Vec<EvaluationTask<'a>>)>);
-
-impl<'a> ScheduleSlice<'a> {
-    // Todo: Impl sort tasks in same deadline by eval order
-    // Todo: Impl remove task from slice
+pub(crate) struct DynamicDeadline {
+    pub(crate) due: Time,
+    pub(crate) tasks: Vec<EvaluationTask>,
 }
 
-pub(crate) struct DynamicSchedule<'a> {
-    queue: PriorityQueue<ScheduledTask<'a>, Reverse<Time>>,
+impl DynamicDeadline {
+    pub(crate) fn sort(&mut self, ir: &RtLolaMir) {
+        self.tasks.sort_by_key(|s| match s {
+            EvaluationTask::Evaluate(idx, _) => ir.outputs[*idx].eval_layer().inner(),
+            EvaluationTask::Spawn(idx) => ir.outputs[*idx].spawn_layer().inner(),
+            EvaluationTask::Close(_, _) => usize::MAX,
+        });
+    }
 }
 
-impl<'a> DynamicSchedule<'a> {
+pub(crate) struct DynamicSchedule {
+    queue: PriorityQueue<ScheduledTask, Reverse<Time>>,
+}
+
+impl DynamicSchedule {
     pub(crate) fn new() -> Self {
         DynamicSchedule { queue: PriorityQueue::new() }
     }
@@ -35,52 +44,58 @@ impl<'a> DynamicSchedule<'a> {
     pub(crate) fn schedule_evaluation(
         &mut self,
         target: OutputReference,
-        parameter: &'a [Value],
+        parameter: &[Value],
         now: Time,
         period: Duration,
     ) {
-        let task = ScheduledTask { task: EvaluationTask::Evaluate(target, parameter), period };
+        let task = ScheduledTask { task: EvaluationTask::Evaluate(target, parameter.to_vec()), period };
         self.queue.push(task, Reverse(now + period));
     }
 
     /// Schedule the close evaluation of a stream or of an instance if parameters are given.
-    pub(crate) fn schedule_close(
-        &mut self,
-        target: OutputReference,
-        parameter: &'a [Value],
-        now: Time,
-        period: Duration,
-    ) {
-        let task = ScheduledTask { task: EvaluationTask::Close(target, parameter), period };
+    pub(crate) fn schedule_close(&mut self, target: OutputReference, parameter: &[Value], now: Time, period: Duration) {
+        let task = ScheduledTask { task: EvaluationTask::Close(target, parameter.to_vec()), period };
         self.queue.push(task, Reverse(now + period));
     }
 
     /// Removes a scheduled evaluation from the schedule
-    pub(crate) fn remove_evaluation(&mut self, target: OutputReference, parameter: &'a [Value], period: Duration) {
-        let task = ScheduledTask { task: EvaluationTask::Evaluate(target, parameter), period };
+    pub(crate) fn remove_evaluation(&mut self, target: OutputReference, parameter: &[Value], period: Duration) {
+        let task = ScheduledTask { task: EvaluationTask::Evaluate(target, parameter.to_vec()), period };
         self.queue.remove(&task);
     }
 
     /// Removes a scheduled close from the schedule
-    pub(crate) fn remove_close(&mut self, target: OutputReference, parameter: &'a [Value], period: Duration) {
-        let task = ScheduledTask { task: EvaluationTask::Close(target, parameter), period };
+    pub(crate) fn remove_close(&mut self, target: OutputReference, parameter: &[Value], period: Duration) {
+        let task = ScheduledTask { task: EvaluationTask::Close(target, parameter.to_vec()), period };
         self.queue.remove(&task);
     }
 
-    /// Returns all scheduled task until and including the given time
-    pub(crate) fn get_schedule_until(&mut self, now: Time) -> ScheduleSlice<'a> {
-        let mut res: Vec<(Time, Vec<EvaluationTask>)> = vec![];
-        while self.queue.peek().is_some() && self.queue.peek().unwrap().1 .0 <= now {
-            let (t, due) = self.queue.pop().unwrap();
-            // Reschedule Task
-            self.queue.push(t.clone(), Reverse(due.0 + t.period));
-            if res.last().is_some() && res.last().unwrap().0 == due.0 {
-                res.last_mut().unwrap().1.push(t.task);
-            } else {
-                res.push((due.0, vec![t.task]));
-            }
+    /// Returns the next scheduled task until and including the given time
+    pub(crate) fn get_next_deadline(&mut self, now: Time) -> Option<DynamicDeadline> {
+        if self.queue.peek().is_none() || self.queue.peek().unwrap().1 .0 > now {
+            return None;
         }
-        ScheduleSlice(res)
+        let (task, task_due) = self.queue.pop().unwrap();
+        // Reschedule Task
+        self.queue.push(task.clone(), Reverse(task_due.0 + task.period));
+
+        let mut deadlines: Vec<EvaluationTask> = vec![task.task];
+        let due = task_due.0;
+
+        // Pop all tasks that are due at the same time
+        while self.queue.peek().is_some() && self.queue.peek().unwrap().1 .0 == due {
+            let (task, task_due) = self.queue.pop().unwrap();
+            // Reschedule Task
+            self.queue.push(task.clone(), Reverse(task_due.0 + task.period));
+            deadlines.push(task.task);
+        }
+
+        Some(DynamicDeadline { due, tasks: deadlines })
+    }
+
+    /// Return the time when the next deadline is due or None if there is no next deadline
+    pub(crate) fn get_next_deadline_due(&self) -> Option<Time> {
+        self.queue.peek().map(|(_, due)| due.0)
     }
 }
 
@@ -88,6 +103,7 @@ impl<'a> DynamicSchedule<'a> {
 mod tests {
     use crate::coordination::dynamic_schedule::DynamicSchedule;
     use crate::coordination::EvaluationTask;
+    use crate::Value;
     use std::time::Duration;
 
     #[test]
@@ -98,29 +114,34 @@ mod tests {
         schedule.schedule_close(1, &[], now, Duration::from_secs(2));
         schedule.schedule_evaluation(2, &[], now, Duration::from_secs(7));
 
-        let res = schedule.get_schedule_until(Duration::from_secs(10));
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(2));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
-        assert_eq!(res.0[0].0, Duration::from_secs(2));
-        assert_eq!(res.0[0].1, vec![EvaluationTask::Close(1, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(4));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
-        assert_eq!(res.0[1].0, Duration::from_secs(4));
-        assert_eq!(res.0[1].1, vec![EvaluationTask::Close(1, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(5));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(0, vec![])]);
 
-        assert_eq!(res.0[2].0, Duration::from_secs(5));
-        assert_eq!(res.0[2].1, vec![EvaluationTask::Evaluate(0, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(6));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
-        assert_eq!(res.0[3].0, Duration::from_secs(6));
-        assert_eq!(res.0[3].1, vec![EvaluationTask::Close(1, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(7));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(2, vec![])]);
 
-        assert_eq!(res.0[4].0, Duration::from_secs(7));
-        assert_eq!(res.0[4].1, vec![EvaluationTask::Evaluate(2, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(8));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
-        assert_eq!(res.0[5].0, Duration::from_secs(8));
-        assert_eq!(res.0[5].1, vec![EvaluationTask::Close(1, &[])]);
-
-        assert_eq!(res.0[6].0, Duration::from_secs(10));
-        assert!(res.0[6].1.contains(&EvaluationTask::Close(1, &[])));
-        assert!(res.0[6].1.contains(&EvaluationTask::Evaluate(0, &[])));
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(10));
+        assert!(res.tasks.contains(&EvaluationTask::Close(1, vec![])));
+        assert!(res.tasks.contains(&EvaluationTask::Evaluate(0, vec![])));
     }
 
     #[test]
@@ -131,24 +152,82 @@ mod tests {
         schedule.schedule_close(1, &[], now, Duration::from_secs(2));
         schedule.schedule_evaluation(2, &[], now, Duration::from_secs(7));
 
-        let res = schedule.get_schedule_until(Duration::from_secs(4));
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(2));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
-        assert_eq!(res.0[0].0, Duration::from_secs(2));
-        assert_eq!(res.0[0].1, vec![EvaluationTask::Close(1, &[])]);
-
-        assert_eq!(res.0[1].0, Duration::from_secs(4));
-        assert_eq!(res.0[1].1, vec![EvaluationTask::Close(1, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(4));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
 
         schedule.remove_close(1, &[], Duration::from_secs(2));
-        let res = schedule.get_schedule_until(Duration::from_secs(10));
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(5));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(0, vec![])]);
 
-        assert_eq!(res.0[0].0, Duration::from_secs(5));
-        assert_eq!(res.0[0].1, vec![EvaluationTask::Evaluate(0, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(7));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(2, vec![])]);
 
-        assert_eq!(res.0[1].0, Duration::from_secs(7));
-        assert_eq!(res.0[1].1, vec![EvaluationTask::Evaluate(2, &[])]);
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(10));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(0, vec![])]);
 
-        assert_eq!(res.0[2].0, Duration::from_secs(10));
-        assert_eq!(res.0[2].1, vec![EvaluationTask::Evaluate(0, &[])]);
+        schedule.remove_evaluation(0, &[], Duration::from_secs(5));
+        let res = schedule.get_next_deadline(Duration::from_secs(20)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(14));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(2, vec![])]);
+        assert!(schedule.get_next_deadline(Duration::from_secs(20)).is_none());
+
+        let res = schedule.get_next_deadline(Duration::from_secs(30)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(21));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(2, vec![])]);
+        schedule.remove_evaluation(2, &[], Duration::from_secs(7));
+        assert!(schedule.get_next_deadline(Duration::from_secs(50)).is_none());
+    }
+
+    #[test]
+    fn test_involved() {
+        let mut schedule = DynamicSchedule::new();
+        let now = Duration::default();
+        schedule.schedule_evaluation(0, &[], now, Duration::from_secs(5));
+        schedule.schedule_close(1, &[], now, Duration::from_secs(2));
+        schedule.schedule_evaluation(2, &[], now, Duration::from_secs(7));
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(2));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(4));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
+
+        let para = vec![Value::Bool(true), Value::Signed(42)];
+        schedule.schedule_evaluation(3, &para, Duration::from_secs(4), Duration::from_secs(1));
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(5));
+        assert!(res.tasks.contains(&EvaluationTask::Evaluate(0, vec![])));
+        assert!(res.tasks.contains(&EvaluationTask::Evaluate(3, para.clone())));
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(6));
+        assert!(res.tasks.contains(&EvaluationTask::Close(1, vec![])));
+        assert!(res.tasks.contains(&EvaluationTask::Evaluate(3, para.clone())));
+
+        schedule.remove_evaluation(3, &para, Duration::from_secs(1));
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(7));
+        assert_eq!(res.tasks, vec![EvaluationTask::Evaluate(2, vec![])]);
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(8));
+        assert_eq!(res.tasks, vec![EvaluationTask::Close(1, vec![])]);
+
+        let res = schedule.get_next_deadline(Duration::from_secs(10)).unwrap();
+        assert_eq!(res.due, Duration::from_secs(10));
+        assert!(res.tasks.contains(&EvaluationTask::Close(1, vec![])));
+        assert!(res.tasks.contains(&EvaluationTask::Evaluate(0, vec![])));
     }
 }