+++ src/coordination/mod.rs
@@ -6,6 +6,7 @@ mod time_driven_manager;
 
 // Re-exports
 pub(crate) use self::controller::Controller;
+pub(crate) use self::dynamic_schedule::DynamicSchedule;
 pub(crate) use self::event_driven_manager::EventEvaluation;
 pub use self::monitor::Monitor;
 pub(crate) use self::time_driven_manager::EvaluationTask;
@@ -16,9 +17,9 @@ use crate::storage::Value;
 pub type Event = Vec<Value>;
 
 #[derive(Debug, Clone)]
-pub(crate) enum WorkItem<'a> {
+pub(crate) enum WorkItem {
     Event(EventEvaluation, Time),
-    Time(TimeEvaluation<'a>, Time),
+    Time(TimeEvaluation, Time),
     End,
 }